// src/app/modules/academics/attendance/model/student-attendance-add-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StudentAttendanceStatusView } from './student-attendance-status-view';


/**
 * StudentAttendenceAddView
 */
export interface StudentAttendanceAddView { 
    /**
     * Gets or Sets StudentAttendanceStatusView
     */
    studentAttendanceStatusViews?: Array<StudentAttendanceStatusView>;
}