// src/app/modules/academics/attendance/model/class-attendance-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassAttendanceView
 */
export interface ClassAttendanceView { 
    /**
     * Gets or sets ClassId
     */
    classId?: string;
    /**
     * Gets or sets SectionId
     */
    sectionId?: string;
    /**
     * Gets or sets CalendarDate
     */
    calendarDate?: Date;
}